Re-use existing sticker picker widgets
There's a potential for a race condition by deleting then recreating the widget. Instead, we'll just re-use the widget and change the properties of it.

Fixes https://github.com/turt2live/matrix-dimension/issues/200

Note that the actual change may not be reflected due to https://github.com/vector-im/riot-web/issues/7547