Add BigBlueButton widget to integration manager
This adds the widget and the configuration for it to the integration
manager, so that the user can add a BBB widget. The code that will
actually run inside of the widget is not yet here.

A few CSS things are updated as well to make them more generic, as
we reused a few things that were previously jitsi only.