Add API backend
This commit adds the join API endpoint that will be used by the widget
to transform a greenlight URL to a BigBlueButton meeting URL.

The full flow is defined within the code itself, but it roughly boils
down to taking a greenlight URL that the user pastes it, sending it to
Dimension, Dimension making some API calls to greenlight to "join" the
meeting and retrieving a join link, before passing that back down to the
client to load.

Unfortunately, while BigBlueButton's server has a nice API, it's useless
to us if all we have is a greenlight link, so we need to do this hacky
route instead.